requirements.txt
PyNaCl
requests
